% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation_effect1.R
\name{beffect}
\alias{beffect}
\title{Title}
\usage{
beffect(
  index,
  u,
  mintercept,
  BB,
  type,
  ocoe_tm,
  ocoe_confounder,
  ocoe_intercept,
  ocoe_mm
)
}
\arguments{
\item{\code{index}}{is the mediator of interest}

\item{\code{u}}{is the state vector}

\item{\code{mintercept}}{the estimated intercept in mediator model}

\item{\code{BB}}{is the estimated state transition matrix}

\item{\code{type}}{the type of outcome}

\item{\code{ocoe_tm}}{the coefficients of treatment-mediator
interaction in outcome model}

\item{\code{ocoe_confounder}}{is the coefficient of confounder
in outcome model}

\item{\code{ocoe_intercept}}{is the estiamted intercept in
outcome model}

\item{\code{ocoe_mm}}{is the coefficient of mediator-mediator interaction
in outcome model}
}
\value{
mediation effects
}
\description{
Title
}
