% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation2.R
\name{binary}
\alias{binary}
\title{Title}
\usage{
binary(
  size = 200,
  tc_m = 0.5,
  mc_m = 0.5,
  tc_o = 0.5,
  mc_o = 0.5,
  sigma_m = 0.5,
  sigma_o = 0.5
)
}
\arguments{
\item{\code{size}}{is sample size}

\item{\code{tc_m}}{coefficient of treatment in mediatior model}

\item{\code{mc_m}}{coefficient of mediator in mediator model}

\item{\code{tc_o}}{coefficient of treatment in outcome model}

\item{\code{mc_o}}{coefficient of mediator in outcome model}

\item{\code{sigma_m}}{variance of error in mediaor model}

\item{\code{sigma_o}}{variance of error in outcome model}
}
\value{
a size-by-12 matrix
}
\description{
Title
}
\examples{
data=binary()
}
