% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ime.R
\name{ime}
\alias{ime}
\title{Main function}
\usage{
ime(index, u, AA, type, form, data)
}
\arguments{
\item{\code{index}}{is the mediator of interest}

\item{\code{u}}{is the state vector}

\item{\code{AA}}{is the  transition matrix}

\item{\code{type}}{the type of outcome}

\item{\code{form}}{specify the regression form for outcome}

\item{\code{data}}{is the n by (p+2+q) data matrix}
}
\value{
mediation effects
}
\description{
Main function
}
