% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOutcome.R
\name{kids}
\alias{kids}
\title{Kids()}
\usage{
kids(BB, index)
}
\arguments{
\item{\code{BB}}{represents the estimated coefficients}

\item{\code{index}}{is the mediator of interest.}
}
\value{

}
\description{
Find the descendants
}
