% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression}
\alias{regression}
\title{Title}
\usage{
regression(form, type, data, C)
}
\arguments{
\item{\code{form}}{is the list having two elements. The first is
a vector representing the cofficients of interaction term between
treatment and mediator. The second element is the coefficients of
interaction term between mediator and mediator.}

\item{\code{type}}{indicate the type of outcome}

\item{\code{data}}{\code{n} by \code{p+2} observation matrix
on treatment, mediators and outcome.}

\item{\code{C}}{are the confounders}
}
\value{
all the estimates of unknown parameters.
}
\description{
Title
}
