% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcSample.R
\name{mcsample}
\alias{mcsample}
\title{Internal function for function \code{beffect()}}
\usage{
mcsample(mcov, mintercept, BB, index, u, size)
}
\arguments{
\item{\code{mcov}}{is the variance matrix of error in
mediator model}

\item{\code{mintercept}is}{the estimated intercept
in mediator model}

\item{\code{BB}}{is the estimated state transition matrix}

\item{\code{index}}{is the mediator of interest}

\item{\code{u}}{is the state vector}

\item{\code{size}}{is the number of observations to simulate.}
}
\value{
A list with length 2. The first element is a size-by-p
matrix representing the observations of mediators with treatment; the second represents
the observations of mediators without treatment.
}
\description{
Internal function for function \code{beffect()}
}
